Move GatewayConfigType to PaymentBundle
